chore : testDataInit -> Application PreLoader로 변경
